+++ ratis-common/src/main/java/org/apache/ratis/util/NettyUtils.java
@@ -45,6 +45,8 @@ import org.slf4j.LoggerFactory;
 
 import javax.net.ssl.KeyManager;
 import javax.net.ssl.TrustManager;
+import java.security.Provider;
+import java.security.Security;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
@@ -191,7 +193,7 @@ public interface NettyUtils {
   }
 
   /**
-   * Apply the {@link SslProvider}, enabled TLS protocols and cipher suites from the given
+   * Apply the {@link SslProvider}, JSSE provider, enabled TLS protocols and cipher suites from the given
    * {@link TlsConf} to the builder, so that the Netty DataStream transport honours the same
    * configuration instead of falling back on the provider defaults.
    *
@@ -205,15 +207,35 @@ public interface NettyUtils {
     if (sslProvider != null) {
       b.sslProvider(sslProvider);
     }
+    final Provider jsseProvider = getJsseProvider(tlsConf);
+    if (jsseProvider != null) {
+      b.sslProvider(SslProvider.JDK).sslContextProvider(jsseProvider);
+    }
     final List<String> protocols = tlsConf.getProtocols();
     if (protocols != null && !protocols.isEmpty()) {
-      b.protocols(protocols.toArray(new String[0]));
+      b.protocols(protocols);
     }
     final List<String> cipherSuites = tlsConf.getCipherSuites();
     if (cipherSuites != null && !cipherSuites.isEmpty()) {
       b.ciphers(cipherSuites, SupportedCipherSuiteFilter.INSTANCE);
     }
     return b;
+  }
+
+  /**
+   * @return the named JSSE {@link Provider} from {@link TlsConf#getJsseProviderName()}, or null when
+   *     unset; throws {@link IllegalArgumentException} when the named provider is not registered.
+   */
+  static Provider getJsseProvider(TlsConf tlsConf) {
+    final String providerName = tlsConf.getJsseProviderName();
+    if (providerName == null || providerName.trim().isEmpty()) {
+      return null;
+    }
+    final Provider namedProvider = Security.getProvider(providerName.trim());
+    if (namedProvider == null) {
+      throw new IllegalArgumentException("JSSE provider not found: " + providerName);
+    }
+    return namedProvider;
   }
 
   static SslContext buildSslContextForClient(TlsConf tlsConf) {
